Make Opacity Support Decimals
The user can edit the save file manually to decimal values for more fine-grain control.

+++ Common/Settings.cs
@@ -18,7 +18,7 @@ namespace SamsidParty_TopNotify
 
 
         // From 0 To 5 (0 Is Fully Opaque, 5 Is Mostly Transparent)
-        public byte Opacity = 0;
+        public float Opacity = 0;
 
         public static Settings Get()
         {
